requirements.txt
Flask
PyPDF2
firebase-admin
google-generativeai
python-dotenv
pdfkit
qrcode
python-docx
Werkzeug
markdown